requirements.txt
langchain_mcp_adapters==0.1.14
openpyxl==3.1.5
akasha-terminal
pandas